+++ lib/latency.js
@@ -1,5 +1,5 @@
 import * as crypto from 'crypto'
-import {showResult} from './show.js'
+import {showResult} from './result.js'
 
 
 /**
