+++ lib/result.js
